Feature (providers): Add initial TV Provider
Added `watchSeries` as a TV Provider.

Added functionality to search for a TV show, given the title, season and
episode.

NOTE: From the Terarium providers for WatchSeries, only 1 of them wasn't
dead. For that reason, there is only 1 provider here currently, which is
bad and needs to be fixed...

Changes in this PR:
- New folder structuer for TV and Movie providers.
- Added TV Provider (WatchSeries).
- Added new route for searching for TV shows: `/api/search/tv`